requirements.txt
streamlit==1.22.0
pandas==1.5.3
psycopg2-binary==2.9.6
sqlalchemy==2.0.15
langchain==0.0.184
openai==0.27.7
python-dotenv==1.0.0
openpyxl==3.1.2